refactor: Extract classes HeaderList and Header from TOC
This is to avoid a loose array of [int, string] pairs. Instead the
header level and text are represented by class Header.

Meanwhile HeaderList keeps track of the header text and how often it
occurs. When a Header gets added, it receives a unique counter.
This way duplicate anchor links are avoided by adding a "-1", "-2", etc.