test: Test Markdown in PHPDocs does not confuse DocBlock
This new test builds upon the existing DocBlock test from
phpdocumentor/reflection-docblock. For this to work, the parent class
needed to be added to the autoloader.

It is now asserted that PHPDocs written in Markdown do not break the
parser. Mainly this is about using `@placeholder@` in comments:
ReflectionDocBlock always assumes `^@` marks a PHPDoc tag.

Instead it must be allowed to use fenced code or indented blocks to
mention writeme placeholders.